CalcWorks

+++ Calculator/src/main/java/com/example/Calculator/Dto/CalculatorDto.java
@@ -13,5 +13,5 @@ public class CalculatorDto {
 
     private long number2;
 
-    private long answer;
+    private double answer;
 }
